Moving helper functions to utils folder
