+++ frontend/src/components/report_pet/PetReport.js
@@ -2,8 +2,8 @@ import React, { useContext, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { AuthContext } from '../../contexts/AuthContext';
 import { AppStateContext } from '../../contexts/AppStateContext';
-import { handleError, clearError } from '../HelperFunctions.js';
-import { isFieldRequired } from '../HelperFunctions.js';
+import { handleError, clearError } from '../../utils/HelperFunctions.js';
+import { isFieldRequired } from '../../utils/HelperFunctions.js';
 import PetReportOptionalData from './PetReportOptionalData';
 import { ReactComponent as ArrowDown } from '../../assets/icons/togglearrow.svg';
 import ImageUpload from './ImageUpload';
