+++ frontend/src/utils/HelperFunctions.js
